pyproject.toml
[tool.poetry]
name = "jarvis"
version = "0.1.0"
description = ""
authors = ["Kristian H.T. <kristian@kristianht.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
